Updated plugin version to 1.6 (multiple changes)
Added wjwwood's serial library for serial port communication. Changed
the way avrdude command is executed on Windows, so that a file
containing avrdude's output is generated.  Created function openPort and
exposed it to JavaScript, which allows implicitelly opening a port from
Javascript console. Added debugging function (enableDebug and
disableDebug) which are also exposed to Javascript.