Added debugging verbosity levels
Debug level 1
This level prints out only basic information such as the avrdude command and the output stored in the output file.
Debug level 2
This level prints out all messages that level 1 prints as well as some more information for uploading and communication with the serial ports.
Debug level 3
This level prints out all messages that the 2 previous levels print and also a message for almost every function that is called during program execution. The output of this level is stored in a .txt file that can be found in the plugin folder.